requirements.txt
opencv_python==4.11.0.86
pyserial==3.5
PySide6==6.8.2.1
PySide6==6.8.3
PySide6_Addons==6.8.2.1
PySide6_Essentials==6.8.2.1
Requests==2.32.3
